🎨 Bedre lagdeling og logging

+++ oving04/backend/src/main/java/edu/jakobkg/calculatorbackend/models/CalculationResult.java
@@ -1,3 +1,3 @@
 package edu.jakobkg.calculatorbackend.models;
 
-public record CalculationResult(Float result, String status) {}
+public record CalculationResult(Double result, String status) {}
